Precommit test

+++ src/screener/data/client.py
@@ -1,9 +1,9 @@
-import requests
+from decimal import Decimal
 from pydantic import BaseModel
 
 class Quote(BaseModel):
     symbol: str
-    price: int
+    price: Decimal 
 
 def retrieve_symbol_data(symbol: str, start_date: str, end_date: str) -> Quote:
